CDAP-5239: Move SecurityRequestContext to cdap-security-spi
