+++ cdap-app-fabric/src/main/java/co/cask/cdap/gateway/handlers/AuthorizationHandler.java
@@ -20,7 +20,6 @@ import co.cask.cdap.common.BadRequestException;
 import co.cask.cdap.common.FeatureDisabledException;
 import co.cask.cdap.common.conf.CConfiguration;
 import co.cask.cdap.common.conf.Constants;
-import co.cask.cdap.common.http.SecurityRequestContext;
 import co.cask.cdap.common.logging.AuditLogEntry;
 import co.cask.cdap.gateway.handlers.util.AbstractAppFabricHttpHandler;
 import co.cask.cdap.proto.security.Action;
@@ -28,8 +27,10 @@ import co.cask.cdap.proto.security.AuthorizationRequest;
 import co.cask.cdap.proto.security.CheckAuthorizedRequest;
 import co.cask.cdap.proto.security.GrantRequest;
 import co.cask.cdap.proto.security.RevokeRequest;
+import co.cask.cdap.security.spi.authentication.SecurityRequestContext;
 import co.cask.cdap.security.spi.authorization.Authorizer;
 import co.cask.http.HttpResponder;
+import com.google.common.base.Objects;
 import com.google.common.base.Optional;
 import com.google.inject.Inject;
 import org.jboss.netty.handler.codec.http.HttpRequest;
@@ -65,8 +66,8 @@ public class AuthorizationHandler extends AbstractAppFabricHttpHandler {
                               HttpResponseStatus responseStatus) throws UnknownHostException {
     String reqBody = String.format("[%s %s %s]", request.getPrincipal(), request.getEntity(), request.getActions());
     AuditLogEntry logEntry = new AuditLogEntry();
-    logEntry.setUserName(SecurityRequestContext.getUserId().or("-"));
-    logEntry.setClientIP(InetAddress.getByName(SecurityRequestContext.getUserIP().or("0.0.0.0")));
+    logEntry.setUserName(Objects.firstNonNull(SecurityRequestContext.getUserId(), "-"));
+    logEntry.setClientIP(InetAddress.getByName(Objects.firstNonNull(SecurityRequestContext.getUserIP(), "0.0.0.0")));
     logEntry.setRequestLine(httpRequest.getMethod(), httpRequest.getUri(), httpRequest.getProtocolVersion());
     logEntry.setRequestBody(reqBody);
     logEntry.setResponseCode(responseStatus.getCode());
