+++ cdap-data-fabric/src/main/java/co/cask/cdap/data2/audit/KafkaAuditPublisher.java
@@ -18,12 +18,13 @@ package co.cask.cdap.data2.audit;
 
 import co.cask.cdap.common.conf.CConfiguration;
 import co.cask.cdap.common.conf.Constants;
-import co.cask.cdap.common.http.SecurityRequestContext;
 import co.cask.cdap.proto.audit.AuditMessage;
 import co.cask.cdap.proto.audit.AuditPayload;
 import co.cask.cdap.proto.audit.AuditType;
 import co.cask.cdap.proto.id.EntityId;
+import co.cask.cdap.security.spi.authentication.SecurityRequestContext;
 import com.google.common.base.Charsets;
+import com.google.common.base.Objects;
 import com.google.common.base.Supplier;
 import com.google.common.base.Suppliers;
 import com.google.gson.Gson;
@@ -59,7 +60,7 @@ public class KafkaAuditPublisher implements AuditPublisher {
 
   @Override
   public void publish(EntityId entityId, AuditType auditType, AuditPayload auditPayload) {
-    String userId = SecurityRequestContext.getUserId().or("");
+    String userId = Objects.firstNonNull(SecurityRequestContext.getUserId(), "");
     AuditMessage auditMessage = new AuditMessage(System.currentTimeMillis(), entityId, userId, auditType, auditPayload);
     LOG.trace("Publishing audit message {}", auditMessage);
